refactor(session): isolate agent environment builder
